+++ src/main/java/org/acumos/microservice/services/impl/GenerateMicroserviceController.java
@@ -492,11 +492,11 @@ public class GenerateMicroserviceController extends DockerizeModel implements Do
 								mData = null;
 							} catch (AcumosServiceException e) {
 								mData = null;
+								HttpStatus httpCode = HttpStatus.INTERNAL_SERVER_ERROR;
 								MDC.put(OnboardingLogConstants.MDCs.RESPONSE_STATUS_CODE,
 										OnboardingLogConstants.ResponseStatus.ERROR.name());
-								MDC.put(OnboardingLogConstants.MDCs.RESPONSE_DESCRIPTION, e.getMessage());
+								MDC.put(OnboardingLogConstants.MDCs.RESPONSE_DESCRIPTION, httpCode.toString());
 								logger.error( "RevertbackOnboarding Failed");
-								HttpStatus httpCode = HttpStatus.INTERNAL_SERVER_ERROR;
 								return new ResponseEntity<ServiceResponse>(
 										ServiceResponse.errorResponse(e.getErrorCode(), e.getMessage()), httpCode);
 							}
@@ -527,11 +527,11 @@ public class GenerateMicroserviceController extends DockerizeModel implements Do
 				HttpStatus httpCode = HttpStatus.INTERNAL_SERVER_ERROR;
 				MDC.put(OnboardingLogConstants.MDCs.RESPONSE_STATUS_CODE,
 						OnboardingLogConstants.ResponseStatus.ERROR.name());
-				MDC.put(OnboardingLogConstants.MDCs.RESPONSE_DESCRIPTION, e.getMessage());
 				logger.error( e.getErrorCode() + "  " + e.getMessage());
 				if (e.getErrorCode().equalsIgnoreCase(OnboardingConstants.INVALID_PARAMETER)) {
 					httpCode = HttpStatus.BAD_REQUEST;
 				}
+				MDC.put(OnboardingLogConstants.MDCs.RESPONSE_DESCRIPTION, httpCode.toString());
 				return new ResponseEntity<ServiceResponse>(
 						ServiceResponse.errorResponse(e.getErrorCode(), e.getMessage()), httpCode);
 			} catch (HttpClientErrorException e) {